move appendix to standoc: https://github.com/metanorma/metanorma-standoc/issues/987
